+++ json_serializable/lib/src/type_helpers/iterable_helper.dart
@@ -10,16 +10,21 @@ class IterableHelper extends TypeHelper {
 
   @override
   String serialize(DartType targetType, String expression, bool nullable,
-      TypeHelperGenerator serializeNested) {
+      SerializeContext context) {
     if (!_coreIterableChecker.isAssignableFromType(targetType)) {
       return null;
     }
+
+    var args = typeArgumentsOf(targetType, _coreIterableChecker);
+    assert(args.length == 1);
+
+    var keyType = args[0];
 
     // This block will yield a regular list, which works fine for JSON
     // Although it's possible that child elements may be marked unsafe
 
     var isList = _coreListChecker.isAssignableFromType(targetType);
-    var subFieldValue = serializeNested(
+    var subFieldValue = context.serialize(
         _getIterableGenericType(targetType), _closureArg, nullable);
 
     var optionalQuestion = nullable ? '?' : '';
@@ -28,6 +33,15 @@ class IterableHelper extends TypeHelper {
     // will be identical to `substitute` – so no explicit mapping is needed.
     // If they are not equal, then we to write out the substitution.
     if (subFieldValue != _closureArg) {
+      if (context.useWrappers && isList) {
+        var method = '\$wrapList';
+        if (nullable) {
+          method = '${method}HandleNull';
+        }
+
+        return '$method<$keyType>($expression, ($_closureArg) => $subFieldValue)';
+      }
+
       // TODO: the type could be imported from a library with a prefix!
       expression =
           '${expression}${optionalQuestion}.map(($_closureArg) => $subFieldValue)';
@@ -47,7 +61,7 @@ class IterableHelper extends TypeHelper {
 
   @override
   String deserialize(DartType targetType, String expression, bool nullable,
-      TypeHelperGenerator deserializeNested) {
+      DeserializeContext context) {
     if (!_coreIterableChecker.isAssignableFromType(targetType)) {
       return null;
     }
@@ -55,7 +69,7 @@ class IterableHelper extends TypeHelper {
     var iterableGenericType = _getIterableGenericType(targetType);
 
     var itemSubVal =
-        deserializeNested(iterableGenericType, _closureArg, nullable);
+        context.deserialize(iterableGenericType, _closureArg, nullable);
 
     // If `itemSubVal` is the same, then we don't need to do anything fancy
     if (_closureArg == itemSubVal) {