+++ json_serializable/CHANGELOG.md
@@ -6,6 +6,15 @@
 * **BREAKING** The deprecated `annotations.dart` library has been removed.
   Use `package:json_annotation` instead.
 
+* **BREAKING** The arguments to `TypeHelper` `serialize` and `deserialize` have
+  changed.
+  * `SerializeContext` and `DeserializeContext` (new classes) are now passed 
+    instead of the `TypeHelperGenerator` typedef (which has been deleted).
+
+* `JsonSerializableGenerator` now supports an optional `useWrappers` argument
+  when generates and uses wrapper classes to (hopefully) improve the speed and
+  memory usage of serialization – at the cost of more code. 
+
 * Make `null` field handling smarter. If a field is classified as not
   `nullable`, then use this knowledge when generating serialization code –  even
   if `includeIfNull` is `false`.